backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
mongoengine==0.28.0
python-dotenv==1.0.0
Werkzeug==3.0.1
pymongo==4.7.2
certifi==2024.2.2
dnspython==2.4.2
numpy==1.26.4
pandas==2.1.3
scikit-learn==1.3.2
pennylane==0.40.0
torch==2.3.1